refactor(broker): step 5/6 — finalize directory split with orchestrator + testing seam
Final step of decomposing replay-check.ts. Two coordinated changes:

1. `replay-check.ts` → `replay-check/index.ts`
   The 446-LOC file that remained after the comparator extractions is
   now the directory's index module. It owns runReplayCheck (the scan
   loop) and re-exports the public surface
   (ReplayCheckReport / ReplayDiscrepancy / ReplayedBudget +
   BudgetCandidateIndexes for type-inference callers). Sibling imports
   simplify from `./replay-check/foo.ts` to `./foo.ts`.

   `cost-ledger/index.ts` and `cost-ledger/routes.ts` consumers, plus
   the budget-candidate-index and unsafe-lifetime-accumulator modules
   that previously type-imported `../replay-check.ts`, all point at
   `./replay-check/index.ts` / `./discrepancy.ts` instead.

2. New `replay-check/testing.ts` (50 LOC)
   The `__replayCheckTesting` frozen object and the
   `__createBudgetCandidateIndexesForTesting` re-export now live in
   their own module. Keeps the public index focused on
   `runReplayCheck` only, signals "this is the test seam" by file
   name, and lets tests import from a stable explicit module path
   instead of fishing test helpers out of the orchestrator.

   `tests/cost-ledger.spec.ts` updated to import from
   `../src/cost-ledger/replay-check/testing.ts`.

Side cleanup: removed imports of internals that the orchestrator no
longer references directly (`addBudgetToIndex`, `removeBudgetFromIndex`,
`crossesThresholdBigInt`, `MAX_*_BIG`, `arraysEqual`, `AuditEventKind`,
`splitAgentDayKey`, `parseStoredThresholds`, `ComputeExpectedCrossingsArgs`)
since the comparator modules own those call sites now.

Final structure
  cost-ledger/replay-check/
  ├── index.ts                       387 LOC  orchestrator (runReplayCheck)
  ├── discrepancy.ts                 374 LOC  ReplayDiscrepancy union + state
  ├── crossing-comparators.ts        341 LOC  threshold-crossing comparators
  ├── threshold-oracle.ts            163 LOC  independent oracle
  ├── internal.ts                    126 LOC  DB rows + key encoders + helpers
  ├── budget-candidate-index.ts      123 LOC  scope-keyed candidate index
  ├── budget-comparator.ts           100 LOC  cost_budgets vs replayed
  ├── aggregate-comparators.ts        85 LOC  cost_by_agent / cost_by_task
  ├── unsafe-lifetime-accumulator.ts  76 LOC  boundary-signal helpers
  └── testing.ts                      50 LOC  __replayCheckTesting

All 249 broker tests pass; typecheck clean; biome clean; file-size hook
OK (largest file 387 LOC, well under the 1500 ceiling).
